backend/requirements.txt
fastapi>=0.88.0
uvicorn>=0.25.0
sqlalchemy>=2.0.0
pymongo>=4.4.1
motor>=3.2.0
pydantic[email]>=2.5.2
python-dotenv==1.0.0
psycopg2-binary
alembic>=1.9.1
asyncpg>=0.27.0
python-multipart>=0.0.6
google-generativeai>=0.3.0
pydantic-settings
tenacity>=8.2.2
redis>=4.5.5
bcrypt>=3.2.0
python-jose>=3.3.0
httpx>=0.24.0
aiofiles>=23.1.0  # For async file operations